+++ OEW-main/src/screens/workstation/components/TimelineRulerGrid.tsx
@@ -1,5 +1,4 @@
-/** @jsx React.createElement */
-import React, { useContext, useLayoutEffect, useRef, useState } from "react";
+import { useContext, useLayoutEffect, useRef, useState } from "react";
 import { PreferencesContext, WorkstationContext } from "../../../contexts";
 import { TimelinePosition, SnapGridSizeOption } from "../../../services/types/types";
 import { formatDuration, measureSeconds } from "../../../services/utils/general";