+++ OEW-main/src/screens/workstation/components/TrackVolumeSlider.tsx
@@ -1,4 +1,4 @@
-import React, { CSSProperties, useContext, useEffect, useMemo, useState } from "react";
+import { CSSProperties, useContext, useEffect, useMemo, useState } from "react";
 import { AutomationLaneEnvelope, Track } from "../../../services/types/types";
 import { formatVolume, normalizedToVolume, volumeToNormalized } from "../../../services/utils/utils";
 import Slider from "../../../components/widgets/Slider";